add profile image upload

+++ controllers/task.js
@@ -1,7 +1,8 @@
 const TaskModel = require("../models/task.js");
 const mongoose = require("mongoose");
+const Joi = require("joi");
 
-const gettasks = async (req, res) => {
+const getTasks = async (req, res) => {
   const { userId } = req;
 
   try {
@@ -13,7 +14,7 @@ const gettasks = async (req, res) => {
   }
 };
 
-const gettask = async (req, res) => {
+const getTask = async (req, res) => {
   const { id } = req.params;
 
   // check if requested user is also owner of task
@@ -33,9 +34,22 @@ const gettask = async (req, res) => {
   }
 };
 
-const createtask = async (req, res) => {
+const createTask = async (req, res) => {
   const task = req.body;
+  //validate
+  const taskschema = Joi.object().keys({
+    taskname: Joi.string().required(),
+    description: Joi.string().required(),
+    date: Joi.string().required(),
+    starttime: Joi.string().required(),
+    endtime: Joi.string().required(),
+  });
 
+  try {
+    await taskschema.validateAsync(task);
+  } catch (error) {
+    return res.status(400).send(error);
+  }
   const newtask = new TaskModel({
     ...task,
     userId: req.userId,
@@ -51,43 +65,91 @@ const createtask = async (req, res) => {
   }
 };
 
-const updatetask = async (req, res) => {
+const updateTask = async (req, res) => {
   const { id } = req.params;
+  const { userId } = req;
   const { taskname, description, date, starttime, endtime } = req.body;
 
   if (!mongoose.Types.ObjectId.isValid(id))
     return res.status(404).send(`No task with id: ${id}`);
 
-  const updatedtask = {
+  const updatedTask = {
     taskname,
     description,
     date,
     starttime,
     endtime,
-    userId: req.userId,
+    userId,
     _id: id,
   };
 
-  await TaskModel.findByIdAndUpdate(id, updatedtask, { new: true });
+  //validate
+  const taskschema = Joi.object().keys({
+    taskname: Joi.string().required(),
+    description: Joi.string().required(),
+    date: Joi.string().required(),
+    starttime: Joi.string().required(),
+    endtime: Joi.string().required(),
+  });
 
-  res.json(updatedtask);
+  try {
+    await taskschema.validateAsync({
+      taskname,
+      description,
+      date,
+      starttime,
+      endtime,
+    });
+  } catch (error) {
+    return res.status(400).send(error);
+  }
+
+  try {
+    const task = await TaskModel.findById(id);
+    if (task.userId != userId) {
+      res.status(404).json({ message: "Invalid request" });
+    } else {
+      try {
+        await TaskModel.findByIdAndUpdate(id, updatedTask, { new: true });
+        res.status(200).json(updatedTask);
+      } catch (error) {
+        res.status(404).json({ message: error.message });
+      }
+    }
+  } catch (error) {
+    res.status(404).json({ message: error.message });
+  }
+
+  // res.json(updatedtask);
 };
 
-const deletetask = async (req, res) => {
+const deleteTask = async (req, res) => {
   const { id } = req.params;
+  const { userId } = req;
 
   if (!mongoose.Types.ObjectId.isValid(id))
     return res.status(404).send(`No task with id: ${id}`);
-
-  await TaskModel.findByIdAndRemove(id);
-
-  res.json({ message: "task deleted successfully." });
+  try {
+    const task = await TaskModel.findById(id);
+    if (task.userId != userId) {
+      res.status(404).json({ message: "Invalid request" });
+    } else {
+      try {
+        await TaskModel.findByIdAndRemove(id);
+        res.status(200).json({ message: "task deleted successfully." });
+      } catch (error) {
+        res.status(404).json({ message: error.message });
+      }
+    }
+  } catch (error) {
+    res.status(404).json({ message: error.message });
+  }
 };
 
 module.exports = {
-  gettasks,
-  gettask,
-  createtask,
-  updatetask,
-  deletetask,
+  getTasks,
+  getTask,
+  createTask,
+  updateTask,
+  deleteTask,
 };
